Expenses working!
git-svn-id: http://radek.unfuddle.com/svn/radek_budget@61 138173b9-53cd-42c6-b595-420ba7eea4ed

+++ models/loans.py
@@ -127,7 +127,7 @@ class LoansBase():
 
             loan_id = s.object_id
 
-        LoansTable.query.filter(LoansTable.id == loan_id).delete()
+        LoansTable.query.filter(and_(LoansTable.user == self.user_id, LoansTable.id == loan_id)).delete()
 
         SlugsTable.query\
         .filter(and_(SlugsTable.object_id == loan_id, SlugsTable.type == 'loan', SlugsTable.user == self.user_id))\
